apps/api-app/requirements.txt
Flask==3.0.3
requests==2.32.3
opentelemetry-api==1.28.0
opentelemetry-sdk==1.28.0
opentelemetry-exporter-otlp==1.28.0
opentelemetry-instrumentation-flask==0.49b0
opentelemetry-instrumentation-requests==0.49b0
opentelemetry-instrumentation-logging==0.49b0

